split off utils.array_utils and utils.file_utils from utils.utils
