+++ docs/source/lpne.utils.rst
@@ -9,10 +9,28 @@ lpne.utils package
 Submodules
 ----------
 
+lpne.utils.array_utils module
+-----------------------------
+
+.. automodule:: lpne.utils.array_utils
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
 lpne.utils.data module
 ----------------------
 
 .. automodule:: lpne.utils.data
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+lpne.utils.file_utils module
+----------------------------
+
+.. automodule:: lpne.utils.file_utils
     :members:
     :undoc-members:
     :show-inheritance:
